Updated types and usage
