+++ apps/apple-settings/app.json
@@ -1,7 +1,6 @@
 {
   "expo": {
     "name": "apple-settings",
-    "icon": "https://icogen.vercel.app/api/icon?icon=🔌&color_hex=1832BA",
-    "plugins": ["./settings.plugin.js"]
+    "icon": "https://icogen.vercel.app/api/icon?icon=🔌&color_hex=1832BA"
   }
 }
